+++ pydantic/fields.py
@@ -6,8 +6,8 @@ from . import errors as errors_
 from .class_validators import Validator, ValidatorSignature, get_validator_signature
 from .error_wrappers import ErrorWrapper
 from .types import Json, JsonWrapper
-from .utils import display_as_type, lenient_issubclass, list_like
-from .validators import NoneType, dict_validator, find_validators, is_none_validator
+from .utils import Callable, ForwardRef, display_as_type, lenient_issubclass, list_like
+from .validators import NoneType, dict_validator, find_validators
 
 Required: Any = Ellipsis
 
@@ -118,6 +118,11 @@ class Field:
         if self.type_ is None:
             raise errors_.ConfigError(f'unable to infer type for attribute "{self.name}"')
 
+        if type(self.type_) == ForwardRef:
+            # self.type_ is currently a ForwardRef and there's nothing we can do now,
+            # user will need to call model.update_forward_refs()
+            return
+
         self.validate_always: bool = (
             getattr(self.type_, 'validate_always', False) or any(v.always for v in self.class_validators.values())
         )
@@ -140,6 +145,8 @@ class Field:
         origin = getattr(self.type_, '__origin__', None)
         if origin is None:
             # field is not "typing" object eg. Union, Dict, List etc.
+            return
+        if origin is Callable:
             return
         if origin is Union:
             types_ = []
@@ -351,7 +358,7 @@ class Field:
         """
         False if this is a simple field just allowing None as used in Unions/Optional.
         """
-        return len(self.validators) > 1 or self.validators[0][1] != is_none_validator
+        return self.type_ != NoneType
 
     def is_complex(self):
         """