+++ HISTORY.rst
@@ -7,6 +7,7 @@ v0.19.0 (unreleased)
 ....................
 * Support ``Callable`` type hint, fix #279 by @proofit404
 * Fix schema for fields with ``validator`` decorator, fix #375 by @tiangolo
+* fix for 'missing' error with ``validate_all`` or ``validate_always``, #381 by @samuelcolvin
 
 v0.18.2 (2019-01-22)
 ....................