+++ HISTORY.rst
@@ -7,6 +7,8 @@ v0.19.0 (unreleased)
 ....................
 * Support ``Callable`` type hint, fix #279 by @proofit404
 * Fix schema for fields with ``validator`` decorator, fix #375 by @tiangolo
+* Add ``multiple_of`` constraint to ``ConstrainedDecimal``, ``ConstrainedFloat``, ``ConstrainedInt``
+  and their related types ``condecimal``, ``confloat``, and ``conint`` #371, thanks @StephenBrown2
 
 v0.18.2 (2019-01-22)
 ....................